fix warning in g++: size() returns size_t, not int

+++ day01/sonar_sweep.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 int num_larger(vector<int> &a) {
     int cnt = 0;
-    for (int i = 1; i < a.size(); i++)
+    for (size_t i = 1; i < a.size(); i++)
         if (a[i] > a[i-1])
             cnt++;
     return cnt;
@@ -32,7 +32,7 @@ int main(int argc, char *argv[]) {
 
     // compute sliding windows
     vector<int> windows;
-    for (int i = 0; i < depth.size() - 2; i++)
+    for (size_t i = 0; i < depth.size() - 2; i++)
         windows.push_back(depth[i] + depth[i+1] + depth[i+2]);
     cout << "Part 2: " << num_larger(windows) << endl;
 }
